feat(package/spotify): show track/album
